fix: exercise_11 hydra/launcher issue
Added a while loop in run.py because it seems that when using hydra/launcher it happens that the artifact containing
the training data is not read properly by pd.read_csv. There might be some issues with the way hydra/launcher handles parallel
tasks by using all CPU cores available. The impact is while the same file is read by all the simultaneous runs, and this file
has been previously preprocessed and does not contain null values, some runs will have a dataframe with null values and an
exception will be raised for them. The solution is to keep reading the file until it is read properly, hence the while loop.